EOS-24299 motr/io_req.c: Ignore service failures when parent node is failed/offline
In device_check() when a node is marked as failed/offline, then don't count the
service failures for checking if a pool version is dud or not.
If the parent node is failed/offline, then only the node should affect the tolerance
of the pool, not the services and devices under the node.

This patch addresses the issue of reduced tolerance of a cluster during degraded
read when a node with multiple controller fails. Each controller within the
failed node is counted towards the max allowed failure at controller level,
which artifically reduces the tolerance of the cluster.

Ex- In a 7 node cluster with N=4, K=3, S=0 with each node having
    3 contollers, and allowance set to [0, 0, 3, 3, 3] can support
    3 node failures, but 1 node failure counts towards 3 service failure.
    This uses up all allowed failures at controller level thus making the
    cluster unable to support any more node failures.

Signed-off-by: Abhishek Saha <abhishek.saha@seagate.com>