lib: introduce m0_trace_level_allow() API
It will allow motr apps to specify the needed level of tracing
allowed in run-time and reduce the impact of the tracing on the
scalability of multi-threaded apps.

Note: `m0_trace_allot()` uses the atomic variable to allocate
the trace records in the buffer, so with many trace records in
each thread of a multithreaded application, the more threads
are working in parallel, the more they get hampered by the
tracing subsystem.

Signed-off-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>